Add the device detail fragment
