+++ app/src/main/res/layout/fragment_device_detail.xml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:padding="16dp">
+
+    <TextView
+        android:id="@+id/device_detail_name"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textAppearance="?textAppearanceHeadline6"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:text="Device name" />
+
+    <TextView
+        android:id="@+id/device_detail_rssi"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textAppearance="?textAppearanceSubtitle1"
+        app:layout_constraintTop_toBottomOf="@id/device_detail_name"
+        tools:text="RSSI: -65" />
+
+    <TextView
+        android:id="@+id/device_detail_last_connected"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textAppearance="?textAppearanceSubtitle1"
+        app:layout_constraintTop_toBottomOf="@id/device_detail_rssi"
+        tools:text="Last Connected: ..." />
+
+    <TextView
+        android:id="@+id/device_detail_connection_seconds"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textAppearance="?textAppearanceCaption"
+        android:visibility="gone"
+        app:layout_constraintTop_toBottomOf="@id/device_detail_last_connected"
+        tools:text="Connected for ... seconds"
+        tools:visibility="visible" />
+
+    <TextView
+        android:id="@+id/device_detail_error"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:background="?colorError"
+        android:gravity="center"
+        android:text="@string/device_detail_error"
+        android:textAppearance="?textAppearanceCaption"
+        android:textColor="?colorOnError"
+        android:visibility="gone"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/device_detail_connection_seconds"
+        app:layout_constraintWidth="wrap_content"
+        tools:visibility="visible" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
